requirements.txt
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
distlib==0.3.3
dnspython==2.1.0
filelock==3.4.0
Flask==2.0.2
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
platformdirs==2.4.0
pymongo==3.12.1
python-dotenv==0.19.2
requests==2.26.0
six==1.16.0
tmdbsimple==2.8.1
typing-extensions==3.10.0.2
urllib3==1.26.7
virtualenv==20.10.0
Werkzeug==2.0.2
zipp==3.6.0
